+++ mondex/dsl/src/main/kotlin/com/anaplan/engineering/azuki/mondex/dsl/action/WorldActions.kt
@@ -1,10 +1,10 @@
 package com.anaplan.engineering.azuki.mondex.dsl.action
 
 interface WorldActions : WorldDeclarableActions {
-    fun createWorld(authPurses: Map<String, Pair<Int, Int>>)
+    fun createWorld(authPurses: Map<String, Pair<ULong, ULong>>)
 }
 
 interface WorldDeclarableActions {
-    fun absTransfer(transferDetails: Triple<String, String, Int>)
+    fun absTransfer(transferDetails: Triple<String, String, ULong>)
     fun absIgnore()
 }
